requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
argilla==1.1.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
backports.zoneinfo==0.2.1
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.14
Deprecated==1.2.14
duckduckgo-search==3.8.5
et-xmlfile==1.1.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.101.1
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.3
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.33.1
gradio_client==0.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.189
latex2mathml==3.76.0
linkify-it-py==2.0.2
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdtex2html==1.2.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.4
olefile==0.46
openai==0.27.10
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.4
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.20
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.27.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
unstructured==0.7.1
urllib3==2.0.4
uvicorn==0.23.2
validators==0.21.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
py-cpuinfo